Fix millis() rollover issue.

+++ SlowSwitcher/src/SlowSwitcher.cpp
@@ -43,7 +43,7 @@ bool SlowSwitcher::update()
     // If it's time to perform an update, ...
     if (StateMachine::update()) {
         // Has dead time elapsed?
-        bool past = ((millis() - switchTime) >= deadTime);
+        bool past = ((long) (millis() - switchTime) >= deadTime);
         if (m_state == eTurningOn && past) {
             // Dead time elapsed, shift from eTurningOn to eOn.
             m_state = eOn;
